Issue #274 Move DefaultConfig to package-level, as there will be multiple implementations
